Refactored code for frontend login files
